<!-- Thank you for contributing! Please make sure that your code changes
are covered with tests. In case of new features or big changes remember
to adjust the documentation.

In case of an existing issue, reference it using one of the following:

closes: #ISSUE
related: #ISSUE

How to write a good git commit message:
http://chris.beams.io/posts/git-commit/
-->

## What
*Describe what the change is solving*

Implement port allocation for EndpointBindings

## How
*Describe the solution*

Use a port bitmap (from ngrok internal monorepo)

Use a static allocation strategy (noted in comments)

Use a fixed port range from 10000-65535 allowing for ~50k EndpointBindings :see_no_evil:

## Breaking Changes
*Are there any breaking changes in this PR?*

+++ api/bindings/v1alpha1/endpointbinding_types.go
@@ -50,9 +50,9 @@ type EndpointBindingSpec struct {
 	// +kubebuilder:validation:Enum=tcp;http;https;tls
 	Scheme string `json:"scheme"`
 
-	// Port is the Service port this Endpoint uses internally to communicate with its pod forwarders
+	// Port is the Service port this Endpoint uses internally to communicate with its Upstream Service
 	// +kubebuilder:validation:Required
-	Port int32 `json:"port"`
+	Port uint16 `json:"port"`
 
 	// EndpointTarget is the target Service that this Endpoint projects
 	// +kubebuilder:validation:Required
@@ -92,7 +92,7 @@ type EndpointTarget struct {
 	// +kubebuilder:validation:Enum=TCP
 	Protocol string `json:"protocol"`
 
-	// Port is the Service targetPort this Endpoint uses for the Pod Forwarders
+	// Port is the Service targetPort this Endpoint's Target Service uses for requests
 	// +kubebuilder:validation:Required
 	Port int32 `json:"port"`
